refactor: 소거 Footer 버튼 다시 sticky 후 tabMenu style prop 으로 수정 및 식당소거 관련 파일 위치 변경
