+++ frontend/src/domains/restaurantExclude/components/actions/ExcludeActionButtons.tsx
@@ -1,11 +1,33 @@
-import ExcludeSubmitButton from '@domains/restaurantExclude/components/actions/ExcludeSubmitButton';
-
 import Button from '@components/actions/Button';
+import Arrow from '@components/assets/icons/Arrow';
 import DoubleArrow from '@components/assets/icons/DoubleArrow';
 
+import { useRestaurantExcludeContext } from '@domains/restaurantExclude/context/RestaurantExcludeProvider';
+
+import { restaurants } from '@apis/restaurant';
+
 import styled from '@emotion/styled';
+import { generateRouterPath } from '@routes/routePath';
+import { useNavigate, useSearchParams } from 'react-router';
 
 function ExcludeActionButtons() {
+  const { selectedRestaurantIds } = useRestaurantExcludeContext();
+
+  const [searchParams] = useSearchParams();
+  const roomCode = searchParams.get('code') ?? '';
+
+  const navigate = useNavigate();
+  const navigateToPrefer = () => {
+    navigate(generateRouterPath.preferRestaurant(roomCode));
+  };
+
+  const submitExcludeRestaurants = () => {
+    restaurants.patch(selectedRestaurantIds);
+    navigateToPrefer();
+  };
+
+  const disabled = selectedRestaurantIds.length === 0;
+
   return (
     <S.ButtonBox>
       <S.RightButtonWrapper>
@@ -16,10 +38,22 @@ function ExcludeActionButtons() {
             rightIcon={
               <DoubleArrow size="sm" direction="right" color="black" />
             }
+            onClick={navigateToPrefer}
           />
         </S.RightButtonBox>
-
-        <ExcludeSubmitButton />
+        <Button
+          text="다음"
+          size="md"
+          rightIcon={
+            <Arrow
+              size="sm"
+              direction="right"
+              color={disabled ? 'gray' : 'white'}
+            />
+          }
+          onClick={submitExcludeRestaurants}
+          disabled={disabled}
+        />
       </S.RightButtonWrapper>
     </S.ButtonBox>
   );
