+++ frontend/src/domains/restaurantExclude/components/restaurantList/RestaurantItem.stories.tsx
@@ -1,4 +1,4 @@
-import { RestaurantExcludeProvider } from '../context/RestaurantExcludeProvider';
+import { RestaurantExcludeProvider } from '../../context/RestaurantExcludeProvider';
 
 import RestaurantItem from './RestaurantItem';
