+++ frontend/src/domains/restaurantExclude/components/restaurantList/RestaurantItem.tsx
