+++ frontend/src/domains/restaurantExclude/components/restaurantList/RestaurantList.tsx
